fix: refactor changelog notification dropdown to a seperate component
