+++ lang/cs.json
@@ -200,6 +200,8 @@
   "button.submit": "Potvrdit",
   "ByUser": "Od {userName}",
   "CancelEdit": "Zrušit úpravu",
+  "ChangelogNotification.firstLine": "We have new stuff for you!",
+  "ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "checkingBrowser": "Your browser is being verified. If this message doesn't disappear, try to move your mouse or to touch your screen for mobile.",
   "clickHere": "Klikněte zde",
   "Clipboard.Copied": "Zkopírováno!",
@@ -1927,8 +1929,6 @@
   "To": "To",
   "TodaysCharge": "Today's charge",
   "Tooltip.PrivateInfo": "This info is private",
-  "TopBar.ChangelogNotification.firstLine": "We have new stuff for you!",
-  "TopBar.ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "TopContributors.Individuals": "Jednotlivci",
   "TopContributors.Organizations": "Organizace",
   "tos": "podmínky služby",
