+++ lang/de.json
@@ -200,6 +200,8 @@
   "button.submit": "Absenden",
   "ByUser": "Von {userName}",
   "CancelEdit": "Bearbeitung abbrechen",
+  "ChangelogNotification.firstLine": "We have new stuff for you!",
+  "ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "checkingBrowser": "Ihr Browser wird überprüft. Wenn diese Nachricht nicht verschwindet, versuchen Sie die Maus zu bewegen oder den Bildschirm auf dem Handy zu berühren.",
   "clickHere": "Klick hier",
   "Clipboard.Copied": "Kopiert!",
@@ -1927,8 +1929,6 @@
   "To": "To",
   "TodaysCharge": "Today's charge",
   "Tooltip.PrivateInfo": "This info is private",
-  "TopBar.ChangelogNotification.firstLine": "We have new stuff for you!",
-  "TopBar.ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "TopContributors.Individuals": "Personen",
   "TopContributors.Organizations": "Organisationen",
   "tos": "Nutzungsbedingungen",
