+++ lang/en.json
@@ -200,6 +200,8 @@
   "button.submit": "Submit",
   "ByUser": "By {userName}",
   "CancelEdit": "Cancel edit",
+  "ChangelogNotification.firstLine": "We have new stuff for you!",
+  "ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "checkingBrowser": "Your browser is being verified. If this message doesn't disappear, try to move your mouse or to touch your screen for mobile.",
   "clickHere": "Click here",
   "Clipboard.Copied": "Copied!",
@@ -1927,8 +1929,6 @@
   "To": "To",
   "TodaysCharge": "Today's charge",
   "Tooltip.PrivateInfo": "This info is private",
-  "TopBar.ChangelogNotification.firstLine": "We have new stuff for you!",
-  "TopBar.ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "TopContributors.Individuals": "Individuals",
   "TopContributors.Organizations": "Organizations",
   "tos": "terms of service",
