+++ lang/es.json
@@ -200,6 +200,8 @@
   "button.submit": "Enviar",
   "ByUser": "Por {userName}",
   "CancelEdit": "Cancelar edición",
+  "ChangelogNotification.firstLine": "We have new stuff for you!",
+  "ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "checkingBrowser": "Tu navegador está siendo verificado. Si este mensaje no desaparece, intenta mover el ratón o tocar la pantalla si estas en un móvil.",
   "clickHere": "Haga clic aquí",
   "Clipboard.Copied": "¡Copiado!",
@@ -1927,8 +1929,6 @@
   "To": "To",
   "TodaysCharge": "Today's charge",
   "Tooltip.PrivateInfo": "Esta información es privada",
-  "TopBar.ChangelogNotification.firstLine": "We have new stuff for you!",
-  "TopBar.ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "TopContributors.Individuals": "Individuos",
   "TopContributors.Organizations": "Organizaciones",
   "tos": "términos de servicio",
