+++ lang/fr.json
@@ -200,6 +200,8 @@
   "button.submit": "Envoyer",
   "ByUser": "Par {userName}",
   "CancelEdit": "Annuler l'édition",
+  "ChangelogNotification.firstLine": "We have new stuff for you!",
+  "ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "checkingBrowser": "Votre navigateur est en cours de vérification. Si ce message ne disparaît pas, essayez de déplacer votre souris ou de toucher votre écran (pour les mobiles).",
   "clickHere": "Cliquez ici",
   "Clipboard.Copied": "Copié",
@@ -1927,8 +1929,6 @@
   "To": "À",
   "TodaysCharge": "Today's charge",
   "Tooltip.PrivateInfo": "Cette information est privée",
-  "TopBar.ChangelogNotification.firstLine": "We have new stuff for you!",
-  "TopBar.ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "TopContributors.Individuals": "Personnes",
   "TopContributors.Organizations": "Organisations",
   "tos": "conditions d'utilisation",
