+++ lang/it.json
@@ -200,6 +200,8 @@
   "button.submit": "Invia",
   "ByUser": "Da {userName}",
   "CancelEdit": "Annulla modifica",
+  "ChangelogNotification.firstLine": "We have new stuff for you!",
+  "ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "checkingBrowser": "Il tuo browser è in fase di verifica. Se questo messaggio non scompare, prova a spostare il mouse o a toccare lo schermo del tuo cellulare.",
   "clickHere": "Clicca qui",
   "Clipboard.Copied": "Copiato!",
@@ -1927,8 +1929,6 @@
   "To": "To",
   "TodaysCharge": "Today's charge",
   "Tooltip.PrivateInfo": "This info is private",
-  "TopBar.ChangelogNotification.firstLine": "We have new stuff for you!",
-  "TopBar.ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "TopContributors.Individuals": "Individuals",
   "TopContributors.Organizations": "Organizzazioni",
   "tos": "terms of service",
