+++ lang/ko.json
@@ -200,6 +200,8 @@
   "button.submit": "제출",
   "ByUser": "By {userName}",
   "CancelEdit": "변경 취소",
+  "ChangelogNotification.firstLine": "We have new stuff for you!",
+  "ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "checkingBrowser": "Your browser is being verified. If this message doesn't disappear, try to move your mouse or to touch your screen for mobile.",
   "clickHere": "더 알아보기",
   "Clipboard.Copied": "복사되었습니다!",
@@ -1927,8 +1929,6 @@
   "To": "수신",
   "TodaysCharge": "Today's charge",
   "Tooltip.PrivateInfo": "This info is private",
-  "TopBar.ChangelogNotification.firstLine": "We have new stuff for you!",
-  "TopBar.ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "TopContributors.Individuals": "개인",
   "TopContributors.Organizations": "조직",
   "tos": "이용 약관",
