+++ lang/pt-BR.json
@@ -200,6 +200,8 @@
   "button.submit": "Enviar",
   "ByUser": "Por {userName}",
   "CancelEdit": "Cancelar edição",
+  "ChangelogNotification.firstLine": "We have new stuff for you!",
+  "ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "checkingBrowser": "Seu navegador está sendo verificado. Se essa mensagem não desaparecer, tente mover seu mouse ou toque na tela de seu smartfone.",
   "clickHere": "Clique aqui",
   "Clipboard.Copied": "Copiado!",
@@ -1927,8 +1929,6 @@
   "To": "Para",
   "TodaysCharge": "Taxa de hoje",
   "Tooltip.PrivateInfo": "Esta informação é privada",
-  "TopBar.ChangelogNotification.firstLine": "We have new stuff for you!",
-  "TopBar.ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "TopContributors.Individuals": "Indivíduos",
   "TopContributors.Organizations": "Organizações",
   "tos": "termos de serviço",
