+++ lang/ru.json
@@ -200,6 +200,8 @@
   "button.submit": "Отправить",
   "ByUser": "От {userName}",
   "CancelEdit": "Отменить изменения",
+  "ChangelogNotification.firstLine": "We have new stuff for you!",
+  "ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "checkingBrowser": "Ваш браузер проверяется. Если это сообщение не исчезло, попробуйте переместить мышку или прикоснуться к экрану телефона.",
   "clickHere": "Нажмите здесь",
   "Clipboard.Copied": "Скопировано!",
@@ -1927,8 +1929,6 @@
   "To": "Куда",
   "TodaysCharge": "Оплата сегодня",
   "Tooltip.PrivateInfo": "Эта информация приватна",
-  "TopBar.ChangelogNotification.firstLine": "We have new stuff for you!",
-  "TopBar.ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "TopContributors.Individuals": "Физические лица",
   "TopContributors.Organizations": "Организации",
   "tos": "условия использования",
