+++ lang/uk.json
@@ -200,6 +200,8 @@
   "button.submit": "Підтвердити",
   "ByUser": "Від {userName}",
   "CancelEdit": "Скасувати зміни",
+  "ChangelogNotification.firstLine": "We have new stuff for you!",
+  "ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "checkingBrowser": "Your browser is being verified. If this message doesn't disappear, try to move your mouse or to touch your screen for mobile.",
   "clickHere": "Натисніть тут",
   "Clipboard.Copied": "Скопійовано!",
@@ -1927,8 +1929,6 @@
   "To": "To",
   "TodaysCharge": "Сьогоднішня плата",
   "Tooltip.PrivateInfo": "Це приватні відомості",
-  "TopBar.ChangelogNotification.firstLine": "We have new stuff for you!",
-  "TopBar.ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "TopContributors.Individuals": "Особи",
   "TopContributors.Organizations": "Організації",
   "tos": "умови користування",
