+++ lang/zh.json
@@ -200,6 +200,8 @@
   "button.submit": "提交",
   "ByUser": "由 {userName}",
   "CancelEdit": "取消编辑",
+  "ChangelogNotification.firstLine": "We have new stuff for you!",
+  "ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "checkingBrowser": "您的浏览器正在验证中。如果此消息没有消失，请尝试移动您的鼠标或触摸您的屏幕以移动设备。",
   "clickHere": "单击此处",
   "Clipboard.Copied": "已复制！",
@@ -1927,8 +1929,6 @@
   "To": "发送至",
   "TodaysCharge": "今日收入",
   "Tooltip.PrivateInfo": "这个信息是私密的",
-  "TopBar.ChangelogNotification.firstLine": "We have new stuff for you!",
-  "TopBar.ChangelogNotification.secondLine": "Click on the {image} to take a look",
   "TopContributors.Individuals": "个人",
   "TopContributors.Organizations": "组织",
   "tos": "服务条款",
